+++ compiler-tests/src/commonTestFixtures/kotlin/dev/zacsweers/metro/compiler/api/GenerateGraphExtensionExtension.kt
@@ -42,8 +42,7 @@ import org.jetbrains.kotlin.fir.toEffectiveVisibility
 import org.jetbrains.kotlin.fir.types.ConeKotlinType
 import org.jetbrains.kotlin.fir.types.FirResolvedTypeRef
 import org.jetbrains.kotlin.fir.types.builder.buildResolvedTypeRef
-import org.jetbrains.kotlin.fir.types.impl.ConeClassLikeTypeImpl
-import org.jetbrains.kotlin.fir.types.toLookupTag
+import org.jetbrains.kotlin.fir.types.constructClassLikeType
 import org.jetbrains.kotlin.name.CallableId
 import org.jetbrains.kotlin.name.ClassId
 import org.jetbrains.kotlin.name.FqName
@@ -271,8 +270,7 @@ internal class GenerateGraphExtensionExtension(session: FirSession) :
             arguments += buildResolvedQualifierCompat(scopeClassId, scopeSymbol, scopeType)
           }
           coneTypeOrNull =
-            ConeClassLikeTypeImpl(
-              StandardClassIds.KClass.toLookupTag(),
+            StandardClassIds.KClass.constructClassLikeType(
               arrayOf(scopeType),
               isMarkedNullable = false,
             )
@@ -297,8 +295,7 @@ internal class GenerateGraphExtensionExtension(session: FirSession) :
             arguments += buildResolvedQualifierCompat(scopeClassId, scopeClassSymbol, scopeType)
           }
           coneTypeOrNull =
-            ConeClassLikeTypeImpl(
-              StandardClassIds.KClass.toLookupTag(),
+            StandardClassIds.KClass.constructClassLikeType(
               arrayOf(scopeType),
               isMarkedNullable = false,
             )